refactor: Remove calculation logic from MainViewModel.kt
Remove calculation logic from MainViewModel.kt. Moved getResult(),
roundMyAnswer(), and prepareExpression() to new file, Evaluate.kt.
Fix on equal click function that was evaluating invalid functions
on multiple clicks. Modify CalculationException.kt to handle
specific, predefined errors only.
