+++ app/src/main/java/com/gigaworks/tech/calculator/util/CalculationException.kt
@@ -2,5 +2,12 @@ package com.gigaworks.tech.calculator.util
 
 
 class CalculationException(
-    val msg: String
-) : ArithmeticException(msg)
+    val msg: CalculationMessage
+) : ArithmeticException(msg.name)
+
+enum class CalculationMessage {
+    INVALID_EXPRESSION,
+    DIVIDE_BY_ZERO,
+    VALUE_TOO_LARGE,
+    DOMAIN_ERROR
+}
